+++ recipes/source.rb
@@ -17,30 +17,24 @@
 # limitations under the License.
 #
 
-include_recipe "build-essential"
+include_recipe 'build-essential'
 
 remote_file "#{Chef::Config[:file_cache_path]}/daemontools.tar.gz" do
   source node['daemontools']['source_url']
   checksum node['daemontools']['source_checksum']
-  owner "root"
-  notifies :run, "bash[install_daemontools]", :immediately
+  owner 'root'
 end
 
-bash "install_daemontools" do
-  user "root"
+directory '/tmp/daemontools' do
+  recursive true
+end
+
+bash 'install_daemontools' do
+  user 'root'
   cwd Chef::Config[:file_cache_path]
-  code <<-EOH
-    (cd /tmp; mkdir daemontools)
-    (tar zxvf daemontools.tar.gz -C /tmp/daemontools --strip-components 2)
+  not_if { ::File.exist?("#{node['daemontools']['bin_dir']}/svc") }
+  code "(tar zxvf daemontools.tar.gz -C /tmp/daemontools --strip-components 2)
     (cd /tmp/daemontools; perl -pi -e 's/extern int errno;/\#include <errno.h>/' src/error.h)
     (cd /tmp/daemontools; package/compile)
-    (cd /tmp/daemontools; mv command/* #{node['daemontools']['bin_dir']})
-    EOH
-  action :nothing
-  notifies :delete, "directory[/tmp/daemontools]", :immediately
-end
-
-directory "/tmp/daemontools" do
-  recursive true
-  action :nothing
+    (cd /tmp/daemontools; mv command/* #{node['daemontools']['bin_dir']})".gsub(/^\s+/, '')
 end